feat: add unique constraint to banned_users_per_streamer table and enhance chat management with ban functionality
